# docs/openapi/velontri_openapi.yaml
openapi: 3.1.0
info:
  title: Velontri API
  version: 1.0.0
  description: |
    Complete API for all 14 Velontri microservices.
    All responses use a standard envelope: { success, message, data, meta }
    All errors use: { success: false, error: { code, message, field }, request_id }
  contact:
    name: Velontri Engineering
    email: dev@velontri.com
servers:
  - url: https://api.velontri.com/api/v1
    description: Production
  - url: https://api.staging.velontri.com/api/v1
    description: Staging
  - url: http://localhost:8001/api/v1
    description: Auth service (local)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT. Access token TTL is 15 minutes.
...
paths:
...
  /crm/customers:
    get:
      tags: [CRM]
      summary: List customers
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Customer list }
  /crm/customers/{customer_id}/notes:
    post:
      tags: [CRM]
      summary: Add a note to a customer
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: customer_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, example: "Customer interested in premium listings" }
      responses:
        '201': { description: Note added }

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────────
...
  /chat/threads/{thread_id}/messages:
    get:
      tags: [Chat]
      summary: Get message history (last 500)
      parameters:
        - { name: thread_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: token,     in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Message history }